added SplashScreen.kt

+++ data/src/main/java/com/lexwilliam/data/OAuthLocalSource.kt
@@ -8,12 +8,12 @@ interface OAuthLocalSource {
     val state: Flow<String?>
     val accessTokenFlow: Flow<String?>
     val refreshTokenFlow: Flow<String>
-    val expiresInFlow: Flow<Int?>
+    val expiresInFlow: Flow<Long?>
 
     suspend fun setCodeVerifier(code: String)
     suspend fun setState(state: String)
     suspend fun setAccessToken(token: String)
     suspend fun setRefreshToken(token: String)
-    suspend fun setExpireIn(time: Int)
+    suspend fun setExpireIn(time: Long)
     suspend fun clearDataStore()
 }
